requirements.txt
numpy==1.24.4
Pillow==10.0.0
pygame==2.5.1
